rename S3 class/file to LookitS3
